interfaces added + unit test

+++ src/main/java/com/safetynet/alerts/controller/FirestationsController.java
@@ -1,16 +1,17 @@
 package com.safetynet.alerts.controller;
 
 import com.safetynet.alerts.model.Firestation;
+import com.safetynet.alerts.model.Person;
 import com.safetynet.alerts.model.StationCoverage;
 import com.safetynet.alerts.service.FirestationService;
-import org.apache.tomcat.jni.Address;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.text.ParseException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @RequestMapping("/firestation")
 @RestController
@@ -23,25 +24,58 @@ public class FirestationsController {
 
     @PostMapping("/add")
     public List<Firestation> addFirestation(@RequestBody Firestation firestation) {
-        logger.info("eee");
-        return firestationService.add(firestation);
+        List<Firestation> empty = new ArrayList<>();
+
+        logger.info("Request = " + firestation );
+        if(!firestationService.add(firestation).isEmpty()) {
+            logger.info("HTTP GET request received, SUCCESS");
+            return firestationService.add(firestation);
+        } else {
+            logger.error("HTTP GET request received, ERROR");
+            return empty;
+        }
     }
 
     @PutMapping("/update")
     public List<Firestation> updateFirestation(@RequestBody Firestation firestation) {
-        logger.info("eee");
-        return firestationService.update(firestation);
+        List<Firestation> empty = new ArrayList<>();
+
+        logger.info("Request = " + firestation );
+        if(!firestationService.update(firestation).isEmpty()) {
+            logger.info("HTTP GET request received, SUCCESS");
+            return firestationService.update(firestation);
+        } else {
+            logger.error("HTTP GET request received, ERROR");
+            return empty;
+        }
     }
 
     @DeleteMapping("/delete")
     public List<Firestation> deleteFirestation(@RequestParam String address) {
-        logger.info("eee");
-        return firestationService.delete(address);
+        List<Firestation> empty = new ArrayList<>();
+
+        logger.info("Request = " + address );
+        if(!firestationService.delete(address).isEmpty()) {
+            logger.info("HTTP GET request received, SUCCESS");
+            return firestationService.delete(address);
+        } else {
+            logger.error("HTTP GET request received, ERROR");
+            return empty;
+        }
     }
 
     @GetMapping("")
-    public StationCoverage getPeoplesCoverageStation(@RequestParam String stationNumber) throws ParseException {
-        logger.info("eee");
-        return firestationService.getPeoplesCoverageStation(stationNumber);
+    public StationCoverage getPeoplesCoverageStation(@RequestParam String stationNumber) {
+        StationCoverage empty = new StationCoverage();
+
+        logger.info("Request = " + stationNumber );
+        Optional<StationCoverage> firestationOptional = Optional.ofNullable(firestationService.getPeoplesCoverageStation(stationNumber));
+        if(firestationOptional.isPresent()) {
+            logger.info("HTTP GET request received, SUCCESS");
+            return firestationService.getPeoplesCoverageStation(stationNumber);
+        } else {
+            logger.error("HTTP GET request received, ERROR");
+            return empty;
+        }
     }
 }
